+++ src/routes/properly-nested/TextField.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+  import type { StringPath } from '$lib/stringPath';
+
   import type { FieldPath, UnwrapEffects } from '$lib';
   import type { SuperForm } from '$lib/client';
   import type { z, AnyZodObject } from 'zod';
@@ -8,7 +10,7 @@
   type T = $$Generic<AnyZodObject>;
 
   export let form: SuperForm<UnwrapEffects<T>, unknown>;
-  export let field: keyof z.infer<T> | FieldPath<z.infer<T>>;
+  export let field: string & StringPath<z.infer<UnwrapEffects<T>>>;
 
   const { path, value, errors, constraints } = formFieldProxy(form, field);
 </script>